pyproject.toml
[tool.poetry]
name = "gnslgbot"
version = "0.1.0"
description = "A Discord bot with AI capabilities"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"
packages = [{include = "bot"}]  # Ensure the bot package is included

[tool.poetry.dependencies]
python = "^3.11"
discord-py = "^2.5.2"
groq = "^0.18.0"
pynacl = "^1.5.0"
python-dotenv = "^1.0.1"
flask = "^3.0.0"  # Added Flask to prevent errors
openai = "^1.69.0"
pytz = "^2025.2"
gtts = "^2.5.4"
pydub = "^0.25.1"
lavalink = "^5.9.0"
wavelink = "^3.4.1"
psycopg2-binary = "^2.9.10"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
